Added an initial, limited version of the zvm.glk module, which provides a ctypes-based foreign function interface to any Glk shared library.
Also added the slightly-modified source code for CheapGlk, a stdio-based Glk library that is used by the glk integration test program tests/glk_test_program.py and the unit and integration test suie tests/glk_tests.py.

Note that the zmv.glk module is not actually complete; it has only been implemented so that the simple Glk test program could function properly, as a proof-of-concept.  Additional functionality and test cases will be added in the near future.

Also added the root SConstruct (this is like the SCons version of a Makefile) that can be used to build cheapglk on MSVC-based Windows machines as well as Unix machines.

As this is my first commit, please let me know about any coding style or documentation issues you have with it (or any other concerns, for that matter).