Added ZInputStream, ZOutputStream, and ZBufferableOutputStream classes and made ZScreen a subclass of ZBufferableOutputStream.
I'm not sure how much I like the idea that the ZScreen is a ZBufferableOutputStream; it does mesh with ZScreen's current design as a class whose behavior depends on the particularities of its current state, but I'm not sure if that design is a good thing or not.  It certainly makes the interface of the class a whole lot simpler than it could be, though--if we want to make it less stateful, we'd probably have to make it use Glk's window model, which would complicate things a lot.